requirements.txt
asgiref==3.3.4
attrs==21.2.0
build==0.4.0
coverage==5.5
deprecation==2.1.0
Django==3.2.13
iniconfig==1.1.1
packaging==20.9
pep517==0.10.0
pluggy==0.13.1
py==1.10.0
PyJWT==2.2.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
